requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
black==24.4.2
celery==5.4.0
certifi==2024.6.2
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
confluent-kafka==2.4.0
coverage==7.5.3
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-channels==0.7.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
iniconfig==2.0.0
kombu==5.3.7
msgpack==1.0.8
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
pika==1.3.2
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.45
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.2.1
pytest-cov==5.0.0
pytest-django==4.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.4
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
